harshit/drone-analysis-app/backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
motor
python-jose[cryptography]
passlib[bcrypt]
pymongo
paho-mqtt
python-dotenv
